requirements.txt
requests>=2.31.0
pydantic>=2.0.0
click>=8.1.0
typing-extensions>=4.5.0
python-dotenv>=1.0.0
pytest>=7.0.0
responses>=0.23.0
